chore: Add eslint and fix linter issues
