+++ jest.config.ts
@@ -1,0 +1,7 @@
+import type { Config } from 'jest'
+
+const config: Config = {
+	testEnvironment: 'jsdom',
+}
+
+export default config
