+++ package.json
@@ -17,6 +17,8 @@
   ],
   "scripts": {
     "build": "rollup --config rollup.config.mjs",
+    "lint": "eslint lib test",
+    "lint:fix": "eslint --fix lib test",
     "check-types": "tsc",
     "dev": "rollup --config rollup.config.mjs --watch",
     "test": "jest --setupFiles '<rootDir>/test/setup.js'",
@@ -48,16 +50,25 @@
     "@babel/core": "^7.22.1",
     "@babel/preset-env": "^7.22.4",
     "@babel/preset-typescript": "^7.21.5",
+    "@jest/globals": "^29.5.0",
+    "@nextcloud/eslint-config": "^8.3.0-beta.2",
     "@nextcloud/event-bus": "^3.1.0",
     "@rollup/plugin-typescript": "^11.1.1",
+    "@types/jasmine": "^4.3.4",
     "babel-jest": "^29.5.0",
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0",
     "rollup": "^3.23.0",
+    "ts-node": "^10.9.1",
     "typescript": "^5.0.4"
   },
   "engines": {
     "node": "^20.0.0",
     "npm": "^9.0.0"
+  },
+  "eslintConfig": {
+    "extends": [
+      "@nextcloud/eslint-config/typescript"
+    ]
   }
 }
