+++ test/setup.js
@@ -1,8 +1,8 @@
 global.OC = {
 	config: {
-        modRewriteWorking: true,
-    },
-    isUserAdmin() {
-        return false
-    },
+		modRewriteWorking: true,
+	},
+	isUserAdmin() {
+		return false
+	},
 }
